Fix firefox select bug
So, after trying to fix it via JS and diving deep into our custom implementation that handle selection
and focus. I discovered that the actual source of the issue is from the default content-editable CSS implementation
for firefox. Somehow the user-select property is not set to text, causing the selection to be broken.

The fix can be implemented by applying a proper user-select property such as:

```
-moz-user-select: text;
```