Make editor not trigger websocket events when previewing and better selection
The page now will have 2 editor instance, the one used for previewing, the one
used for editing. Using one instance is complicated as we need to turn off plugins
and restart connection and reset metadata. When previewing, the instance for
editing will be hidden.

Selection now works accross the page, however, width are now relative
to the user screen as we're using relative padding to set the page width.
Using max-width is not possible as browser will treat anything outside it
as unselectable.

In the future, we should be able to implement our custom selection method, especially
as we're moving to block based editor like notion.