Fix upload space logo changed space
This issue was hard to debug because it is unpredictable. It happens because the active space index
is pointing to another space as the order of space list changes.

The first call to whoami returned spaces in a particular order, after uploading, the order
changed thus the active index might points to a different space.

The fix was to always order spaces alphabetically.